services/intelligence/pyproject.toml
[project]
name = "intelligence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "shubh305", email = "shubham.305singh@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=0.2.2",
    "openai>=2.16.0",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.40.0",
    "fastapi-limiter>=0.1.6",
    "redis>=5.0.1",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
intelligence = "intelligence:main"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
